Troubleshooting: Unable to Add Markers to Map
I'm facing difficulties in adding markers to the map component. Despite my efforts, the markers aren't rendering correctly or as expected. I've reviewed the code, checked for errors, and explored various solutions, but I haven't been able to resolve the issue yet. This commit marks the current state of the code as I continue to work on getting the markers to function as intended.